+++ cohere/responses/custom_model.py
@@ -2,6 +2,8 @@ from dataclasses import dataclass
 from datetime import datetime
 from typing import Any, Dict, Optional
 
+from cohere.utils import JobWithStatus
+
 try:
     from typing import Literal, TypedDict
 except ImportError:
@@ -73,9 +75,10 @@ class HyperParametersInput(TypedDict):
     learning_rate: float
 
 
-class CustomModel(CohereObject):
+class BaseCustomModel(CohereObject, JobWithStatus):
     def __init__(
         self,
+        wait_fn,
         id: str,
         name: str,
         status: CUSTOM_MODEL_STATUS,
@@ -96,10 +99,12 @@ class CustomModel(CohereObject):
         self.base_model = base_model
         self.model_id = model_id
         self.hyperparameters = hyperparameters
+        self._wait_fn = wait_fn
 
     @classmethod
-    def from_dict(cls, data: Dict[str, Any]) -> "CustomModel":
+    def from_dict(cls, data: Dict[str, Any], wait_fn) -> "BaseCustomModel":
         return cls(
+            wait_fn=wait_fn,
             id=data["id"],
             name=data["name"],
             status=data["status"],
@@ -113,6 +118,84 @@ class CustomModel(CohereObject):
             else None,
         )
 
+    def has_terminal_status(self) -> bool:
+        return self.status == "READY"
+
+
+class CustomModel(BaseCustomModel):
+    def wait(
+        self,
+        timeout: Optional[float] = None,
+        interval: float = 60,
+    ) -> "CustomModel":
+        """Wait for custom model job completion.
+
+        Args:
+            timeout (Optional[float], optional): Wait timeout in seconds, if None - there is no limit to the wait time.
+                Defaults to None.
+            interval (float, optional): Wait poll interval in seconds. Defaults to 60.
+
+        Raises:
+            TimeoutError: wait timed out
+
+        Returns:
+            CustomModel: custom model.
+        """
+
+        return self._wait_fn(custom_model_id=self.id, timeout=timeout, interval=interval)
+
+
+class AsyncCustomModel(BaseCustomModel):
+    async def wait(
+        self,
+        timeout: Optional[float] = None,
+        interval: float = 60,
+    ) -> "CustomModel":
+        """Wait for custom model job completion.
+
+        Args:
+            timeout (Optional[float], optional): Wait timeout in seconds, if None - there is no limit to the wait time.
+                Defaults to None.
+            interval (float, optional): Wait poll interval in seconds. Defaults to 60.
+
+        Raises:
+            TimeoutError: wait timed out
+
+        Returns:
+            CustomModel: custom model.
+        """
+
+        return await self._wait_fn(custom_model_id=self.id, timeout=timeout, interval=interval)
+
+
+@dataclass
+class ModelMetric(CohereObject):
+    created_at: datetime
+    step_num: int
+    loss: Optional[float] = None
+    accuracy: Optional[float] = None
+    f1: Optional[float] = None
+    precision: Optional[float] = None
+    recall: Optional[float] = None
+
+    @classmethod
+    def from_dict(cls, data: Dict[str, Any]) -> "ModelMetric":
+        return cls(
+            created_at=_parse_date_with_variable_seconds(data["created_at"]),
+            step_num=data["step_num"],
+            loss=data.get("loss"),
+            accuracy=data.get("accuracy"),
+            f1=data.get("f1"),
+            precision=data.get("precision"),
+            recall=data.get("recall"),
+        )
+
 
 def _parse_date(datetime_string: str) -> datetime:
     return datetime.strptime(datetime_string, "%Y-%m-%dT%H:%M:%S.%f%z")
+
+
+def _parse_date_with_variable_seconds(datetime_string: str) -> datetime:
+    # model metrics timestamp sometimes contains nanoseconds, so we truncate
+    dt_concat = datetime_string[:26] + datetime_string[-1:]
+    return datetime.strptime(dt_concat, "%Y-%m-%dT%H:%M:%S.%f%z")