+++ tests/sync/test_chat.py
@@ -1,10 +1,8 @@
 import unittest
-from typing import List
 
 from utils import get_api_key
 
 import cohere
-from cohere.responses.chat import Mode
 
 API_KEY = get_api_key()
 co = cohere.Client(API_KEY)
@@ -154,7 +152,7 @@ class TestChat(unittest.TestCase):
         )
         self.assertIsInstance(prediction.text, str)
         self.assertIsInstance(prediction.conversation_id, str)
-        self.assertIsNone(prediction.chatlog)
+        self.assertIsNotNone(prediction.chatlog)
         self.assertIn("User: Hey!", prediction.prompt)
         self.assertIn("Chatbot: Hey! How can I help you?", prediction.prompt)
 
@@ -213,34 +211,3 @@ class TestChat(unittest.TestCase):
         for logit_bias in invalid:
             with self.assertRaises(cohere.error.CohereError):
                 _ = co.chat("Yo what up?", logit_bias=logit_bias, max_tokens=5)
-
-    def test_search_query_generation(self):
-        prediction = co.chat("What are the tallest penguins?", mode="search_query_generation")
-        self.assertIsInstance(prediction.is_search_required, bool)
-        self.assertIsInstance(prediction.queries, List)
-        self.assertTrue(prediction.is_search_required)
-        self.assertGreater(len(prediction.queries), 0)
-
-    def test_search_query_generation_with_enum(self):
-        prediction = co.chat("What are the tallest penguins?", mode=Mode.SEARCH_QUERY_GENERATION)
-        self.assertIsInstance(prediction.is_search_required, bool)
-        self.assertIsInstance(prediction.queries, List)
-        self.assertTrue(prediction.is_search_required)
-        self.assertGreater(len(prediction.queries), 0)
-
-    def test_augmented_generation(self):
-        prediction = co.chat(
-            "What are the tallest penguins?",
-            mode="augmented_generation",
-            documents=[
-                {
-                    "title": "Tall penguins",
-                    "snippet": "Emperor penguins are the tallest",
-                    "url": "http://example.com/foo",
-                }
-            ],
-        )
-        self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.citations, List)
-        self.assertGreater(len(prediction.text), 0)
-        self.assertGreater(len(prediction.citations), 0)